requirements.txt
Django==5.1.3
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
gunicorn==23.0.0
django-cors-headers==4.6.0
django-environ>=0.11.2
uuid==1.30
pillow==11.0.0
graphene-django==3.2.2
djangorestframework==3.15.2
django-filter==24.3
Markdown==3.7
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1